requirements.txt
argon2-cffi
flask
flask-migrate
flask-sqlalchemy
